rename settings_overrides to settings_local; update readme
