+++ .gitignore
@@ -21,7 +21,7 @@ server/docs
 *secrets.py
 *settings_secret.py
 *settings_custom.py
-*settings_overrides.py
+*settings_local.py
 
 docker_repo.var
 server/src
