+++ README.md
@@ -24,14 +24,15 @@ Machine, which is required to run Docker on Mac/Windows hosts.
 After you clone the repository locally, there are several configuration steps required to prepare the project.
 
 
-#### Create settings_secret.py and secrets.py
+#### Create settings_secret.py, settings_local.py, and secrets.py
 
-Create `server/portal/settings/settings_secret.py` containing what is in `secret` field in the `Core Portal Settings Secret` entry secured on [UT Stache](https://stache.security.utexas.edu)
+- Create `server/portal/settings/settings_secret.py` containing what is in `secret` field in the `Core Portal Settings Secret` entry secured on [UT Stache](https://stache.security.utexas.edu)
 
-_Note: Update `_PORTAL_PROJECTS_ID_PREFIX` in `settings_secret.py`to
-be something unique (like `local.username.CEP`) so that project creation in your local development does not clash with other development work (as directories and Tapis storage system names will be created using this prefix)._
+- Copy `server/portal/settings/settings_local.example.py` to `server/portal/settings/settings_local.py`
+	 - _Note: Update `PORTAL_PROJECTS_ID_PREFIX` in `settings_local.py` to be something unique (like `USERNAME.CEP`) so that project creation in your local development does not clash with other development work (as directories and Tapis storage system names will be created using this prefix)._
+    - _Note: [Setup ngrok](#setting-up-notifications-locally) and update `WH_BASE_URL` in `settings_local.py` to enable webhook notifications locally._
 
-Copy `server/conf/cms/secrets.sample.py` to `server/conf/cms/secrets.py`
+- Copy `server/conf/cms/secrets.sample.py` to `server/conf/cms/secrets.py`
 
 #### Build the image for the portal's django container:
     make build
@@ -111,7 +112,7 @@ setup_allocations_index(force=True)
 ```
 ngrok http 443
 ```
-2. Then, take the `https` url generated by ngrok and paste it into the `_WH_BASE_URL` setting in `settings_secret.py`
+2. Then, take the `https` url generated by ngrok and paste it into the `WH_BASE_URL` setting in `settings_local.py`
 
 
 ### Setup local access to the portal:
